src/main/java/com/erp/maisPraTi/service/EmailService.java
package com.erp.maisPraTi.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.mail.SimpleMailMessage;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.stereotype.Service;

@Service
public class EmailService {

    @Autowired
    private JavaMailSender mailSender;

    public void sendPasswordResetEmail(String to, String token){
        String link = "http://localhost:5173/resetpassword?token=" + token;
        String subject = "Recuperação de Senha";
        String body = "Clique no link abaixo para redefinir a sua senha:\n" + link;

        SimpleMailMessage message = new SimpleMailMessage();
        message.setTo(to);
        message.setSubject(subject);
        message.setText(body);
        mailSender.send(message);
    }
}
